Enable safer order of operations for payments
Previously, LPS would add the funds to the patron's account then attempt
to deduct that amount from the jamex machine. This commit reverses the
order of operations such that we deduct the amount from the machine,
then make the api call to add those funds to their Libki account.
If that API call fails, the funds are added back to the Jamex machine
before returning the users change.